layers/fileops-essentials-layer/python/requirements.txt
aws-psycopg2==1.3.8
chardet==5.2.0
install==1.3.5
psycopg2-binary==2.9.7
jinja2==3.1.2
requests==2.31.0
urllib3==1.26.16
